split fixtures and tests
