+++ batchflow/tests/conftest.py
@@ -1,0 +1,5 @@
+""" Pytest configuration. """
+# pylint: disable=invalid-name, unused-import
+import pytest
+
+pytest_plugins = ["batchflow.tests.fixtures.dataformat",]
