+++ batchflow/tests/dataformat_test.py
@@ -1,355 +1,45 @@
-""" Tests for data_d_f functionality for models with multiple inputs. """
+""" Tests for data_format functionality.
+Each test follows the same pattern:
+    First of all, 'fake_data' and 'config' are created via 'setup'
+    fixture. These are Dataset instance and dictionary. They serve
+    as placeholder for real data to train model on and simplest possible
+    configuration for 'train_model' action respectively.
+
+    After that, 'config' is changed by passing 'data_format' option
+    to different keys in it. This step is optional as we want
+    to test default behavior aswell.
+
+    Finally, 'test_pipeline' is created via 'get_pipeline' fixture. It is
+    then applied to 'fake_dataset'. 'next_batch' is used to actually run it.
+
+The idea behind testing protocol is that it is possible to create such architecture
+of neural network (i.e. number of layers, kernels, filters and so on) and
+such input data (i.e. size of images and number of channels) that passing
+wrong 'data_format' parameter would inevitably raise ValueError.
+
+For example, if we try to propagate 28x28x3 image through convolutional
+layer when 'kernel_size' equals 5 and 'data_format' is 'channels_first',
+it would result in negative dimension size as (3-5) is less than 0.
+For more imformation about exact architecture check fixtures file.
+
+Name of test functions has following structure:
+    test_<format of data>_<format passed to config>_<how format is passed>.
+    For example, test_last_first_common mean that created data has 'channels_last'
+    format and 'channels_first' is passed to 'config/common/data_format'.
+
+Most of the tests are used against multiple architectures, list of which is
+defined in TEST_MODELS.
+"""
 # pylint:  disable=import-error, wrong-import-position
 # pylint: disable=missing-docstring, redefined-outer-name
-import os
-import warnings
-warnings.filterwarnings("ignore")
-os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
-
 import pytest
-import numpy as np
-import tensorflow as tf
-
-from ..models.tf.base import TFModel
-from batchflow import Pipeline, ImagesBatch, Dataset
-from batchflow import B, V
-from batchflow.models.tf.layers import conv_block
-from ..models.tf import VGG7, ResNet18, Inception_v1
 
 
-AVAILABLE_MODELS = ['resnet', 'inception']
+TEST_MODELS = ['multi_input', 'single_input',
+               'vgg', 'resnet', 'inception']
 
 
-
-
-@pytest.fixture()
-def single_setup():
-    """ Pytest fixture that is used to generate fake dataset and model
-    config with desired format.
-
-    Parameters
-    ----------
-    d_f: {'channels_last', 'channels_first'}
-        desired format of returns.
-
-    Returns
-    -------
-    tuple
-        First element is instance of Dataset.
-        Second element is dict with model description.
-    """
-    def _single_setup(d_f):
-        if d_f == 'channels_last':
-            shape_in = (32, 32, 4)
-            shape_out = (2, 2, 64)
-        elif d_f == 'channels_first':
-            shape_in = (4, 32, 32)
-            shape_out = (64, 2, 2)
-
-        size = 50
-        batch_shape = (size,) + shape_in
-        output = (size,) + shape_out
-        images_array = np.random.random(batch_shape)
-        labels_array = np.random.choice(5, size=output)
-        data = images_array, labels_array
-        fake_dataset = Dataset(index=size,
-                               batch_class=ImagesBatch,
-                               preloaded=data)
-
-        model_config = {'inputs': {'images': {'shape': shape_in},
-                                   'labels': {'classes': 5, 'shape': shape_out}},
-                        'initial_block/inputs': 'images',
-                        'head': {'layout': 'a'},
-                        'body': {'layout': 'cna cna',
-                                 'filters': [6, 64],
-                                 'kernel_size': [11, 7],
-                                 'strides': [3, 1],
-                                 'padding': ['valid']*2},
-                        'loss': 'l1'}
-        return fake_dataset, model_config
-    return _single_setup
-
-
-@pytest.fixture()
-def get_single_pipeline():
-    """ Creates instance of Pipeline that is configured to use given model
-    with passed parameters.
-
-    Parameters
-    ----------
-    current_config : dict
-        Dictionary with parameters of model.
-
-    model_name : TFModel
-        Model class.
-
-    Returns
-    -------
-    Pipeline
-        Test pipeline that consists of initialization of needed variables and
-        given model and preparing for training with given config.
-    """
-
-    class SingleModel(TFModel):
-        pass
-
-    def _get_single_pipeline(current_config):
-        test_pipeline = (Pipeline()
-                         .init_variable('current_loss')
-                         .init_model('dynamic', SingleModel,
-                                     'TestModelSingle', current_config)
-                         .to_array()
-                         .train_model('TestModelSingle',
-                                      fetches='loss',
-                                      images=B('images'),
-                                      labels=B('labels'),
-                                      save_to=V('current_loss'))
-                         )
-        return test_pipeline
-    return _get_single_pipeline
-
-
-
-@pytest.fixture()
-def multi_setup():
-    class MultiBatch(ImagesBatch):
-        components = 'images_1', 'images_2', 'labels'
-
-    def _multi_setup(d_f):
-        if d_f == 'channels_last':
-            shape_in = (32, 32, 4)
-            shape_out = (8, 8, 16)
-        elif d_f == 'channels_first':
-            shape_in = (4, 32, 32)
-            shape_out = (16, 8, 8)
-
-        size = 50
-        batch_shape = (size,) + shape_in
-        output = (size,) + shape_out
-
-        images_array_1 = np.random.random(batch_shape)
-        images_array_2 = np.random.random(batch_shape)
-        labels_array = np.random.choice(5, size=output)
-        data = images_array_1, images_array_2, labels_array
-        fake_dataset = Dataset(index=size,
-                               batch_class=MultiBatch,
-                               preloaded=data)
-
-        model_config = {'inputs': {'images_1': {'shape': shape_in},
-                                   'images_2': {'shape': shape_in},
-                                   'labels': {'classes': 5, 'shape': shape_out}},
-                        'initial_block/inputs': ['images_1', 'images_2'],
-                        'head': {'layout': 'a'},
-                        'loss': 'l1'}
-        return fake_dataset, model_config
-    return _multi_setup
-
-
-@pytest.fixture()
-def get_multi_pipeline():
-
-    class MultiModel(TFModel):
-        @classmethod
-        def default_config(cls):
-            config = TFModel.default_config()
-            config['body/block_1'] = {}
-            config['body/block_2'] = {}
-            return config
-
-        def build_config(self, names=None):
-            config = super().build_config(names)
-            return config
-
-        @classmethod
-        def body(cls, inputs, name='body', **kwargs):
-            kwargs = cls.fill_params('body', **kwargs)
-            with tf.variable_scope(name):
-                input_1, input_2 = inputs
-                x_1 = cls.block_1(input_1, **kwargs)
-                x_2 = cls.block_2(input_2, **kwargs)
-                dct = {**{'data_format': 'channels_last'},
-                       **cls.fill_params('common', **kwargs),
-                       **cls.fill_params('body', **kwargs),
-                       **cls.fill_params('body', **kwargs)['block_1']}
-                axis = cls.channels_axis(dct['data_format'])
-                output = tf.concat([x_1, x_2], axis=axis)
-            return output
-
-        @classmethod
-        def block_1(cls, input_1, **kwargs):
-            kwargs = cls.fill_params('body/block_1', **kwargs)
-            kwargs = {**kwargs, **kwargs['block_1']}
-
-            with tf.variable_scope('block_1'):
-                x_1 = conv_block(input_1, layout='cna',
-                                 filters=8, kernel_size=11,
-                                 strides=3, padding='valid', name='1', **kwargs)
-            return x_1
-
-        @classmethod
-        def block_2(cls, input_2, **kwargs):
-            kwargs = cls.fill_params('body/block_2', **kwargs)
-            kwargs = {**kwargs, **kwargs['block_2']}
-
-            with tf.variable_scope('block_2'):
-                x_2 = conv_block(input_2, layout='cna',
-                                 filters=8, kernel_size=11,
-                                 strides=3, padding='valid', name='2', **kwargs)
-            return x_2
-
-    def _get_multi_pipeline(current_config):
-        test_pipeline = (Pipeline()
-                         .init_variable('current_loss')
-                         .init_model('dynamic', MultiModel,
-                                     'TestModelMulti', current_config)
-                         .to_array(src='images_1', dst='images_1')
-                         .to_array(src='images_2', dst='images_2')
-                         .train_model('TestModelMulti',
-                                      fetches='loss',
-                                      images_1=B('images_1'),
-                                      images_2=B('images_2'),
-                                      labels=B('labels'),
-                                      save_to=V('current_loss'))
-                         )
-        return test_pipeline
-    return _get_multi_pipeline
-
-
-@pytest.fixture()
-def vgg_setup():
-
-    def _vgg_setup(d_f):
-
-        if d_f == 'channels_last':
-            shape_in = (20, 20, 2)
-
-        elif d_f == 'channels_first':
-            shape_in = (2, 20, 20)
-
-        size = 50
-        batch_shape = (size,) + shape_in
-        images_array = np.random.random(batch_shape)
-        labels_array = np.random.choice(10, size=size)
-        data = images_array, labels_array
-        fake_dataset = Dataset(index=size,
-                               batch_class=ImagesBatch,
-                               preloaded=data)
-
-        model_config = {'inputs': {'images': {'shape': shape_in},
-                                   'labels': {'classes': 10}},
-                        'initial_block/inputs': 'images',
-                        'body/block': {'pool_size': 1, 'pool_strides':1},
-                        'common': {'padding': 'valid'}}
-        return fake_dataset, model_config
-    return _vgg_setup
-
-@pytest.fixture()
-def get_vgg_pipeline():
-
-    def _get_vgg_pipeline(model_class, current_config):
-        test_pipeline = (Pipeline()
-                         .init_variable('current_loss')
-                         .init_model('dynamic', model_class,
-                                     'TestModel', current_config)
-                         .to_array()
-                         .train_model('TestModel',
-                                      fetches='loss',
-                                      images=B('images'),
-                                      labels=B('labels'),
-                                      save_to=V('current_loss'))
-                         )
-        return test_pipeline
-    return _get_vgg_pipeline
-
-
-@pytest.fixture()
-def model_setup():
-
-    def _model_setup(d_f):
-        if d_f == 'channels_last':
-            shape_in = (100, 100, 2)
-
-        elif d_f == 'channels_first':
-            shape_in = (2, 100, 100)
-
-        size = 50
-        batch_shape = (size,) + shape_in
-        images_array = np.random.random(batch_shape)
-        labels_array = np.random.choice(10, size=size)
-        data = images_array, labels_array
-        fake_dataset = Dataset(index=size,
-                               batch_class=ImagesBatch,
-                               preloaded=data)
-
-        model_config = {'inputs': {'images': {'shape': shape_in},
-                                   'labels': {'classes': 10}},
-                        'initial_block/inputs': 'images',
-                        'body/block': {'pool_size': 1, 'pool_strides':1},
-                        'initial_block': {'padding': 'valid'}}
-        return fake_dataset, model_config
-    return _model_setup
-
-@pytest.fixture()
-def get_model_pipeline():
-
-    def _get_model_pipeline(model_class, current_config):
-        if current_config.get('body'):
-            if current_config['body'].get('data_format'):
-                current_config['initial_block']['data_format'] = current_config['body']['data_format']
-        test_pipeline = (Pipeline()
-                         .init_variable('current_loss')
-                         .init_model('dynamic', model_class,
-                                     'TestModel', current_config)
-                         .to_array()
-                         .train_model('TestModel',
-                                      fetches='loss',
-                                      images=B('images'),
-                                      labels=B('labels'),
-                                      save_to=V('current_loss'))
-                         )
-        return test_pipeline
-    return _get_model_pipeline
-
-
-@pytest.fixture()
-def setup(single_setup, multi_setup,
-          vgg_setup, model_setup):
-
-    def _setup(mode, d_f):
-        if mode == 'single_input':
-            return single_setup(d_f)
-        if mode == 'multi_input':
-            return multi_setup(d_f)
-        if mode == 'vgg':
-            return vgg_setup(d_f)
-        if mode in AVAILABLE_MODELS:
-            return model_setup(d_f)
-        return None
-
-    return _setup
-
-
-@pytest.fixture
-def get_pipeline(get_single_pipeline, get_multi_pipeline,
-                 get_vgg_pipeline, get_model_pipeline):
-
-    def _get_pipeline(mode, current_config):
-        if mode == 'single_input':
-            return get_single_pipeline(current_config)
-        if mode == 'multi_input':
-            return get_multi_pipeline(current_config)
-        if mode == 'vgg':
-            return get_vgg_pipeline(VGG7, current_config)
-        if mode == 'resnet':
-            return get_model_pipeline(ResNet18, current_config)
-        if mode == 'inception':
-            return get_model_pipeline(Inception_v1, current_config)
-        return None
-
-    return _get_pipeline
-
-# @pytest.mark.parametrize('mode', ['inception'])
-@pytest.mark.parametrize('mode', ['multi_input', 'vgg', 'single_input'] + AVAILABLE_MODELS)
+@pytest.mark.parametrize('mode', TEST_MODELS)
 class TestPlural:
     def test_last_last_default(self, mode, setup, get_pipeline):
         """ Default value for 'data_format' is 'channels_last'. """
@@ -460,7 +150,7 @@ class TestPlural:
 
 def test_first_first_blocks(setup, get_pipeline):
     """ 'data_format' can be passed to a particular block in body,
-    thgough it should be explicitly stated in model definition.
+    though it should be explicitly stated in model definition.
     """
     fake_dataset, config = setup('multi_input', d_f='channels_first')
